+++ example/client/client1/index.js
@@ -1,12 +1,18 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { CatalogProvider } from 'react-component-catalog'
 
 import catalog, { App } from './catalog'
+import { CatalogProvider } from 'react-component-catalog'
 
+/**
+ * Showcase use of CatalogProvider, even nested ones that consume the previous
+ * context (and merge their catalog with the outer catalog)
+ */
 ReactDOM.render(
   <CatalogProvider catalog={catalog}>
-    <App />
+    <CatalogProvider>
+      <App />
+    </CatalogProvider>
   </CatalogProvider>,
   document.getElementById('_root'),
 )