+++ src/components/use-catalog.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+import CatalogContext from './catalog-context'
+
+export default function useCatalog() {
+  return React.useContext(CatalogContext)
+}